refactor: split universal_lane.rs into a module directory
Pure mechanical move: mod.rs keeps the module doc (including the test
hooks) and shared lane plumbing; submodules hold the per-premise-family
template code (sign, bridge) and shared statement building. External
paths preserved via re-exports.

Gates: corpus exports byte-identical to main (including the
content-hashed universal-lane module filenames inside the json export);
item parity name-by-name; lane tests green.
